Update gradle to javafx-gradle-plugin
