+++ SlidingTab/.gradle/2.5/taskArtifacts/cache.properties
