+++ SlidingTab/.nb-gradle/profiles/private/aux-config
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<gradle-project-properties>
+  <!--DO NOT EDIT THIS FILE! - Used by the Gradle plugin of NetBeans.-->
+  <auxiliary>
+    <editor-bookmarks lastBookmarkId="0" xmlns="http://www.netbeans.org/ns/editor-bookmarks/2"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+      <group>
+        <file>file:/D:/Users/marcus/Documents/Workspace/kairos-example/SlidingTab/build.gradle</file>
+      </group>
+    </open-files>
+  </auxiliary>
+</gradle-project-properties>
