+++ SlidingTab/settings.gradle
