+++ SlidingTab/src/main/resources/fxml/app_activity.fxml
@@ -1,0 +1,23 @@
+<?import javafx.scene.layout.VBox?>
+<?import org.kairos.layouts.SlidingTabLayout?>
+<?import org.kairos.layouts.ViewPager?>
+<?import org.kairos.Toolbar?>
+<?import org.kairos.layouts.DrawerLayout?>
+<?import javafx.scene.control.Label?>
+<DrawerLayout xmlns:fx="http://javafx.com/fxml" prefWidth="800" prefHeight="600" fx:id="drawer">
+    <content>
+        <VBox  spacing="8" alignment="TOP_CENTER">
+            <Toolbar fx:id="toolbar">
+                <SlidingTabLayout fx:id="tabLayout" />
+            </Toolbar>
+            <ViewPager fx:id="viewPager" VBox.vgrow="ALWAYS"/>
+        </VBox>
+    </content>
+    <nav>
+        <VBox style="-fx-background-color: white;">
+            <Label text="This is the nav"/>
+        </VBox>
+    </nav>
+
+
+</DrawerLayout>
