+++ SlidingTab/src/main/resources/fxml/app_fragment.fxml
@@ -1,0 +1,9 @@
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Label?>
+
+<fx:root xmlns:fx="http://javafx.com/fxml" xmlns="http://javafx.com/javafx/8" type="org.kairos.core.Fragment">
+    <VBox AnchorPane.topAnchor="0" AnchorPane.rightAnchor="0"
+          AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" style="-fx-padding:16px">
+        <Label fx:id="label"/>
+    </VBox>
+</fx:root>
