--- src/main/java/Main.java
@@ -1,33 +1,0 @@
-/**
- * Created by marconi on 28/01/16.
- */
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.layout.StackPane;
-import javafx.stage.Stage;
-import org.kairos.components.MaterialButton;
-import org.kairos.core.ActivityFactory;
-
-public class Main extends Application {
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) {
-        StackPane root=new StackPane(); // Create root pane
-        stage.setScene(new Scene(root)); // Set the scene in the stage
-
-        // this object represent the stack  of activities  in your application
-        ActivityFactory factory=new ActivityFactory(stage);
-
-        // set the material design style in your application
-        stage.getScene().getStylesheets().add(MaterialButton.class.getResource("controls.css").toExternalForm());
-        stage.getScene().getStylesheets().add(getClass().getResource("styles.css").toExternalForm());
-
-        factory.startActivity(MainActivity.class); // start the activity
-        stage.show();
-    }
-}
